Update patch for 3.11.5.7
Fixes include adjustments for SessionLeakTest and SchemaChangesCCTest.
First one is adjusted for shard aware cases which have different
expected number of connections in contrast to Cassandra clusters.
Second one has number of nodes increased. This solves the issue of
queries quietly timing out due to cluster not having a quorum for raft.