Switch to running integration-test phase for 4.x versions
Previously the matrix would run integration testing through surefire
plugin during the `test` phase instead of `integration-test`.
Surefire plugin is generally considered to be made for unit testing and
not integration testing.
Such setup makes it skip configuration included in pom.xml for integration
testing that is tied to the failsafe plugin.

In order to return to proper setup this change switches to running
`integration-test` phase which is handled by maven-failsafe-plugin.

This includes a few adjustments which are:
The reports directory is now `failsafe-reports` and not `surefire-reports` (4.x),
the test report pattern now includes the `TEST-` prefix which skips unexpected
summary files for the groups of tests,
`-Dtest='<ignore list>'` was replaced with failsafe plugin appropriate
`-Dit.test='<ignores list>'` with default initial list being empty string
rather than `'*'`.